requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework>=2.2
djangorestframework-jwt
daphne